+++ pkgs/checks/test/extensions/core_test.dart
@@ -14,6 +14,12 @@ void main() {
 
       check(1).isRejectedBy((it) => it.isA<String>(), which: ['Is a int']);
     });
+    test('isNotA', () {
+      check(1).isNotA<String>();
+
+      check(1).isRejectedBy((it) => it.isNotA<int>(), which: ['is a int']);
+      check(1).isRejectedBy((it) => it.isNotA<num>(), which: ['is a num']);
+    });
   });
   group('HasField', () {
     test('has', () {