+++ pkgs/test/CHANGELOG.md
@@ -1,5 +1,14 @@
-## 1.31.2-wip
+## 1.32.0-wip
 
+* Add support for `DART_TEST_REPORTER` environment variable in test runner and
+  when tests are run directly on platforms which support `dart:io`. The
+  environment variable takes precedence over configuration in `dart_test.yaml`
+  but is overridden by the `--reporter` flag when passed to the test runner.
+
+## 1.31.2
+
+* Add support for running tests as native CLI bundles (vm platform only).
+  * You can run tests this way with `--compiler cli`.
 * **Impacts Configuration** Support using the OS platform selector to configure
   browser tests.
   Previously tests loaded for the browser would have an operating system of
@@ -10,7 +19,12 @@
   independent from any OS specific configuration. For instance a wide skip of
   all tests with OS `'windows'` would previously still run browser tests on
   windows, but will now skip all tests including browser tests.
-* Treat calls to `exit` as test failures in VM tests.
+* Use a DevTools URL instead of a defunct observatory URL.
+* Disable throttling in chrome launch arguments.
+* Allow package_config `3.x.x`.
+* Require `analyzer: '>=13.0.0 <15.0.0'`
+* Use the compact or failures-only reporters by default for tests run directly
+  instead of through the test runner.
 
 ## 1.31.1