+++ pkgs/test_core/CHANGELOG.md
@@ -1,7 +1,17 @@
-## 0.6.19-wip
+## 0.6.20-wip
 
+* Add support for `DART_TEST_REPORTER` environment variable.
+
+## 0.6.19
+
+* Add support for `-c cli` (the native CLI compiler) to the vm platform.
 * Support using the OS platform selector to configure browser tests.
-* Treat calls to `exit` as test failures in VM tests.
+* Use a DevTools URL instead of a defunct observatory URL.
+* Allow package_config `3.x.x`.
+* Require `analyzer: '>=13.0.0 <15.0.0'`
+* Update `parse_metadata.dart` to be compatible with `analyzer >=13.0.0 <15.0.0`.
+* Use the compact or failures-only reporters by default for tests run directly
+  instead of through the test runner.
 
 ## 0.6.18